Update files for pagination. Enhance abstract http service
