--- common/src/main/scala/com/github/scalalab3/logs/common/offset/Index.scala
@@ -1,5 +1,0 @@
-package com.github.scalalab3.logs.common.offset
-
-import com.github.scalalab3.logs.common.util.Keys
-
-case class Index(name: String = Keys.time)
