--- common/src/main/scala/com/github/scalalab3/logs/common/offset/OrderBy.scala
@@ -1,8 +1,0 @@
-package com.github.scalalab3.logs.common.offset
-
-case class OrderBy(index: Index = Index(), ordering: Ordering = Desc)
-
-sealed trait Ordering
-
-case object Desc extends Ordering
-case object Asc extends Ordering
