--- common/src/main/scala/com/github/scalalab3/logs/common/offset/Slice.scala
@@ -1,3 +1,0 @@
-package com.github.scalalab3.logs.common.offset
-
-case class Slice(orderBy: OrderBy = OrderBy(), offset: Offset)
