+++ src/main/scala/com/github/scalalab3/logs/services/Messages.scala
@@ -6,6 +6,7 @@ sealed trait AbstractResponse
 
 case class BadRequest(error: String) extends AbstractResponse
 case class LogsResponse(logs: Seq[Log]) extends AbstractResponse
+case class PageLogsResponse(logs: Seq[Log]) extends AbstractResponse
 
 case class Request(query: Option[String])
 case class Page(number: Int, size: Int)
