+++ src/main/scala/com/github/scalalab3/logs/services/read/ReadServiceActor.scala
@@ -1,14 +1,15 @@
 package com.github.scalalab3.logs.services.read
 
 import akka.actor.{Actor, ActorRef}
-import com.github.scalalab3.logs.common.offset.Offset._
-import com.github.scalalab3.logs.common.offset.{Offset, OffsetBound, Slice}
-import com.github.scalalab3.logs.services.Page
+import com.github.scalalab3.logs.common.Offset._
+import com.github.scalalab3.logs.common._
+import com.github.scalalab3.logs.services.{BadRequest, Page}
 
 class ReadServiceActor(dbService: ActorRef) extends Actor {
 
   override def receive: Receive = {
-    case page @ Page(_, _) => dbService forward Slice(offset = bounds(page))
+    case page @ Page(num, size) if num > 0 && size > 0 => dbService forward Slice(offset = bounds(page))
+    case _ => sender ! BadRequest("Bad request")
   }
 
   def bounds(page: Page): Offset = {
